Make BMP085 support optional
My DFRobot Arduino got stuck in setup() with BMP085 support enabled.
Probably it's not so happy with the signals it gets
on the I2C interface. The pins are not in the
"official" order for the Pedelec controller.

To re-enable BPM085 support, just uncomment
the SUPPORT_BMP085 define.

Code size also went down 2.5KB without BMP085 code.