Make battery charge detection configurable
Battery charge detection behaves flaky if the battery
is quite empty: If you switch off the bike for a few hours
and come back, the voltage is quite often above the two volts threshold.

For LiFePo A123 cells the voltage drop occurs only
at the end when the battery is reeeaaly empty.
-> Manual reset is best here anyway.