Add memory debugging code
We recently ran out of SRAM several times and
this results in pretty funny behaviors.

If you enable the new define, you get the amount
of free RAM printed on the serial output.

Also add an own copy of the F() macro to place
strings into flash instead of flash AND program memory.