Always call save_eeprom() when shutting down on purpose
We have "fallback logic" in the code to detect a supply voltage drop
and then quickly write the EEPROM values. The supply voltage is of course
gone when we switch off the FET...

If you connect too many peripherals to the 5V line,
the capacitor to drive the Arudino is emptied too fast
and the EEPROM values will not be saved.

This is a fix for shutdown via the menu,
the S-LCD and J-LCD displays.