Reset PID integral on brake config-option
Use RESET_PID_ON_BRAKE to reset the pid regulator if you are braking or
stop pedaling. If this config option is not active, the pid regulator
will slowly ramp down instead, which gives you a faster kick-in of the
motor if you release the brake or start pedaling again.